niva/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.22
pydantic==1.10.12
motor==3.1.1
python-dotenv==1.0.1
langchain==0.0.300  # check latest compatible version
ollama      # or openai-compatible client
requests==2.31.0
pytest==7.4.0
pandas
pymongo
sentence-transformers
pinecone-client
tqdm
python-multipart